Fix: fixed calc files

+++ src/CalculatorTests.py
@@ -49,8 +49,8 @@ class MyTestCase(unittest.TestCase):
         test_data = CsvReader('/src/square_root.csv').data
         pprint(test_data)
         for row in test_data:
-         self.assertEqual(self.calculator.square_root(row['Value 1']), int(row['Result']))
-         self.assertEqual(self.calculator.result, int(row['Result']))
+         self.assertEqual(self.calculator.square_root(row['Value 1']), float(row['Result']))
+         self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_results_property(self):
         self.assertEqual(self.calculator.result, 0)
